requirements.txt
requests
colorama
libeospy
pytz